SIMKDOWN: Fix DIAGNOSTICS decryptionKey update output overflow
queryAnswer was declared to be int instead of int64_t! Fix.

+++ my-own-problems/simple-markdown-editor/simple-markdown-editor.cpp
@@ -424,7 +424,7 @@ int64_t solveBruteForce(const vector<Query>& queries, vector<string>& bruteForce
                     const int queryPosition = (query.encryptedArgument ^ decryptionKey) - 1;
                     assert(document[queryPosition] == 'X');
                     //cout << "IsRangeFormatted at " << queryPosition << endl;
-                    int queryAnswer = -1;
+                    int64_t queryAnswer = -1;
                     {
                         int openingFormatPos = -1;
                         for (int pos = 0; pos < document.size(); pos++)
